COS-6122:Bulk upload issues fix
