+++ packages/apps/frontera/src/routes/settings/src/components/Tabs/panels/BillingPanel/BillingPanel.tsx
@@ -17,12 +17,12 @@ import {
 import { BankTransferSelectionContextProvider } from '@settings/components/Tabs/panels/BillingPanel/context/BankTransferSelectionContext';
 
 import { cn } from '@ui/utils/cn';
+import { validateEmail } from '@utils/email';
 import { Button } from '@ui/form/Button/Button';
 import { IconButton } from '@ui/form/IconButton';
 import { useStore } from '@shared/hooks/useStore';
 import { DotsVertical } from '@ui/media/icons/DotsVertical';
 import { SlashOctagon } from '@ui/media/icons/SlashOctagon';
-import { validateEmail } from '@shared/util/emailValidation';
 import { useDisclosure } from '@ui/utils/hooks/useDisclosure';
 import { getGraphQLClient } from '@shared/util/getGraphQLClient';
 import { Menu, MenuItem, MenuList, MenuButton } from '@ui/overlay/Menu/Menu';
