+++ packages/apps/frontera/src/routes/src/components/EmailMultiCreatableSelect/MultiValueWithActionMenu.tsx
@@ -10,11 +10,11 @@ import { useLocalStorage } from 'usehooks-ts';
 import { cn } from '@ui/utils/cn';
 import { DataSource } from '@graphql/types';
 import { Input } from '@ui/form/Input/Input';
+import { validateEmail } from '@utils/email';
 import { SelectOption } from '@ui/utils/types';
 import { Edit03 } from '@ui/media/icons/Edit03';
 import { Copy01 } from '@ui/media/icons/Copy01';
 import { useStore } from '@shared/hooks/useStore';
-import { validateEmail } from '@shared/util/emailValidation';
 import { useCopyToClipboard } from '@shared/hooks/useCopyToClipboard';
 import { Menu, MenuItem, MenuList, MenuButton } from '@ui/overlay/Menu/Menu';
 import { useContactCardMeta } from '@organization/state/ContactCardMeta.atom';
